feat(elements): Make connections and icons common
