+++ packages/elements/examples/nextjs/app/sign-up/[[...sign-up]]/page.tsx
@@ -1,7 +1,7 @@
 'use client';
 
-import { GlobalError, Loading } from '@clerk/elements/common';
-import { Action, Provider, ProviderIcon, SignUp, Step, Strategy } from '@clerk/elements/sign-up';
+import { Connection, GlobalError, Icon, Loading } from '@clerk/elements/common';
+import { Action, SignUp, Step, Strategy } from '@clerk/elements/sign-up';
 import type { ComponentProps } from 'react';
 
 import { H1, HR as Hr } from '@/components/design';
@@ -28,29 +28,29 @@ export default function SignUpPage() {
             <H1>Sign Up</H1>
 
             <div className='flex flex-col items-stretch justify-center gap-2'>
-              <Provider
+              <Connection
                 name='github'
                 className='flex items-center justify-center gap-4 text-white rounded bg-[#171717] px-4 py-3 text-sm shadow-sm ring-1 ring-black/[0.06] transition-all hover:bg-opacity-80'
               >
-                <ProviderIcon className='invert' />
+                <Icon className='invert' />
                 Sign In with GitHub
-              </Provider>
+              </Connection>
 
-              <Provider
+              <Connection
                 name='google'
                 className='flex items-center justify-center gap-4 text-white rounded bg-[#333f61] px-4 py-3 text-sm shadow-sm ring-1 ring-black/[0.06] transition-all hover:bg-opacity-80'
               >
-                <ProviderIcon />
+                <Icon />
                 Sign In with Google
-              </Provider>
+              </Connection>
 
-              <Provider
+              <Connection
                 name='metamask'
                 className='flex items-center justify-center gap-4 text-[#161616] rounded bg-white px-4 py-3 text-sm shadow-sm ring-1 ring-black/[0.06] transition-all hover:bg-opacity-80'
               >
-                <ProviderIcon />
+                <Icon />
                 Sign In with Metamask
-              </Provider>
+              </Connection>
             </div>
 
             <Hr />
@@ -91,29 +91,29 @@ export default function SignUpPage() {
             <H1>Sign Up</H1>
 
             <div className='flex flex-col items-stretch justify-center gap-2'>
-              <Provider
+              <Connection
                 name='github'
                 className='flex items-center justify-center gap-4 text-white rounded bg-[#171717] px-4 py-3 text-sm shadow-sm ring-1 ring-black/[0.06] transition-all hover:bg-opacity-80'
               >
-                <ProviderIcon className='invert' />
+                <Icon className='invert' />
                 Sign In with GitHub
-              </Provider>
+              </Connection>
 
-              <Provider
+              <Connection
                 name='google'
                 className='flex items-center justify-center gap-4 text-white rounded bg-[#333f61] px-4 py-3 text-sm shadow-sm ring-1 ring-black/[0.06] transition-all hover:bg-opacity-80'
               >
-                <ProviderIcon />
+                <Icon />
                 Sign In with Google
-              </Provider>
+              </Connection>
 
-              <Provider
+              <Connection
                 name='metamask'
                 className='flex items-center justify-center gap-4 text-[#161616] rounded bg-white px-4 py-3 text-sm shadow-sm ring-1 ring-black/[0.06] transition-all hover:bg-opacity-80'
               >
-                <ProviderIcon />
+                <Icon />
                 Sign In with Metamask
-              </Provider>
+              </Connection>
             </div>
 
             <Hr />
