+++ packages/elements/src/react/common/index.ts
@@ -2,7 +2,8 @@
 import 'client-only';
 
 export { Field, FieldError, FieldState, GlobalError, Input, Label, Submit } from '~/react/common/form';
-export { ProviderIcon } from '~/react/common/providers';
+export { Connection, Icon } from '~/react/common/connections';
+export { Loading } from '~/react/common/loading';
 
 export type {
   FormFieldErrorProps,
@@ -14,7 +15,5 @@ export type {
   FormProps,
   FormSubmitProps,
 } from '~/react/common/form';
-
+export type { ConnectionProps, IconProps } from '~/react/common/connections';
 export type { OTPInputSegmentStatus } from '~/react/common/form/otp';
-
-export { Loading } from '~/react/common/loading';
